# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "HelmetWash API",
        "description": "REST API HelmetWash yang sinkron dengan model layanan, outlet, pesanan, dan riwayat status pada database.",
        "contact": {
            "name": "HelmetWash Support",
            "email": "admin@helmetwash.test"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "HelmetWash API Server"
        }
    ],
    "paths": {
...
        "/api/admin/reports": {
            "get": {
                "tags": [
                    "Reports"
                ],
                "summary": "Laporan dan statistik (Admin)",
                "description": "Mengembalikan ringkasan laporan berdasarkan data pesanan yang ada di database.",
                "operationId": "getReports",
                "responses": {
                    "200": {
                        "description": "Data laporan berhasil didapat",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "summary": {
                                            "properties": {
                                                "totalOrders": {
                                                    "type": "integer",
                                                    "example": 150
                                                },
                                                "totalRevenue": {
                                                    "type": "integer",
                                                    "example": 7500000
                                                },
                                                "avgOrder": {
                                                    "type": "integer",
                                                    "example": 50000
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "byService": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "count": {
                                                        "type": "integer",
                                                        "example": 80
                                                    },
                                                    "revenue": {
                                                        "type": "integer",
                                                        "example": 4000000
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "Paket Premium"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "byDelivery": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "string",
                                                        "example": "pickup"
                                                    },
                                                    "count": {
                                                        "type": "integer",
                                                        "example": 70
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "byStatus": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "status": {
                                                        "type": "string",
                                                        "example": "Selesai"
                                                    },
                                                    "count": {
                                                        "type": "integer",
                                                        "example": 120
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "addonUsage": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "string",
                                                        "example": "antijamur"
                                                    },
                                                    "count": {
                                                        "type": "integer",
                                                        "example": 45
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "byDate": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "date": {
                                                        "type": "string",
                                                        "format": "date",
                                                        "example": "2026-07-02"
                                                    },
                                                    "revenue": {
                                                        "type": "integer",
                                                        "example": 350000
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "orders": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Order"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Tidak terautentikasi"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "Order": {
                "properties": {
                    "id": {
                        "type": "string",
                        "example": "ORD-12345"
                    },
                    "customerName": {
                        "type": "string",
                        "example": "Budi Santoso"
                    },
                    "phone": {
                        "type": "string",
                        "example": "08123456789"
                    },
                    "serviceId": {
                        "type": "string",
                        "example": "basic"
                    },
                    "serviceName": {
                        "type": "string",
                        "example": "Paket Basic"
                    },
                    "addons": {
                        "type": "array",
                        "items": {
                            "type": "string",
                            "example": "antijamur"
                        }
                    },
                    "deliveryId": {
                        "type": "string",
                        "example": "pickup"
                    },
                    "deliveryName": {
                        "type": "string",
                        "example": "Pick-up & Delivery"
                    },
                    "address": {
                        "type": "string",
                        "example": "Jl. Melati No. 5"
                    },
                    "priceBase": {
                        "type": "integer",
                        "example": 25000
                    },
                    "priceAddons": {
                        "type": "integer",
                        "example": 5000
                    },
                    "priceDelivery": {
                        "type": "integer",
                        "example": 10000
                    },
                    "totalPrice": {
                        "type": "integer",
                        "example": 40000
                    },
                    "status": {
                        "type": "string",
                        "example": "Diterima"
                    },
                    "date": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2026-07-02T10:30:00.000Z"
                    },
                    "payment": {
                        "properties": {
                            "method": {
                                "type": "string",
                                "example": "qris"
                            },
                            "status": {
                                "type": "string",
                                "example": "paid"
                            }
                        },
                        "type": "object"
                    },
                    "history": {
                        "type": "array",
                        "items": {
                            "properties": {
                                "status": {
                                    "type": "string",
                                    "example": "Diterima"
                                },
                                "date": {
                                    "type": "string",
                                    "format": "date-time",
                                    "example": "2026-07-02T10:30:00.000Z"
                                }
                            },
                            "type": "object"
                        }
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Masukkan token JWT yang didapat dari POST /api/login",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}